moved few function to cache
